chore : 알림 권한 허용 Home 컨테이너로 이동
