+++ containers/WebRTC/WebRTCContainer.tsx
@@ -1,15 +1,8 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/no-explicit-any */
-/* eslint-disable no-restricted-globals */
 /* eslint-disable @typescript-eslint/naming-convention */
-import { initializeApp } from 'firebase/app';
-import { getMessaging, getToken, onMessage } from 'firebase/messaging';
-import { onBackgroundMessage } from 'firebase/messaging/sw';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
 import { useRecoilValue } from 'recoil';
 
-import { createToken } from '@/apis/alarm';
 import { getCoreEndtime } from '@/apis/coretimes';
 import {
   useChatHandler,
@@ -19,20 +12,16 @@ import {
 } from '@/hooks/Socket';
 import useModal from '@/hooks/useModal';
 import useToggle from '@/hooks/useToggle';
-import { memberIdState } from '@/recoil/atom';
+import { captureTimeState, memberIdState } from '@/recoil/atom';
 import { getNickName } from '@/utils/getNicknames';
 
 import WebRTCLayout from './WebRTCLayout';
 
-declare global {
-  interface Window {
-    registration: any;
-  }
-}
 const WebRTCContainer = () => {
   const router = useRouter();
   const { room_id } = router.query;
   const name = useRecoilValue(memberIdState);
+  const isCaptureTime = useRecoilValue(captureTimeState);
 
   const [curName, setCurName] = useState<string>();
   const [curNickname, setCurNickname] = useState('');
@@ -43,8 +32,6 @@ const WebRTCContainer = () => {
   const {
     produce,
     curMembers,
-    // curDevice,
-    // curPeerList,
     videoStreams,
     audioStreams,
     chattingList,
@@ -63,8 +50,6 @@ const WebRTCContainer = () => {
     addChattingList,
   );
   const coreIssue = useModal();
-
-  const [isCaptureTime, setIsCaptureTime] = useState(false);
 
   const setCoreEndTime = async () => {
     const coreEndTime = await getCoreEndtime(Number(curRoomId));
@@ -104,70 +89,6 @@ const WebRTCContainer = () => {
     handleMike();
   };
 
-  const memberId = useRecoilValue(memberIdState);
-
-  const saveToken = async (token: string) => {
-    await createToken(memberId, token);
-  };
-
-  const askPermission = async () => {
-    const permission = await window.Notification.requestPermission();
-    if (permission !== 'granted') return;
-
-    const firebaseApp = initializeApp({
-      apiKey: 'AIzaSyDjK6isLBGownY7C1AEA6n05-hjpZEleEo',
-      authDomain: 'learniverse-b34d9.firebaseapp.com',
-      projectId: 'learniverse-b34d9',
-      storageBucket: 'learniverse-b34d9.appspot.com',
-      messagingSenderId: '605501909741',
-      appId: '1:605501909741:web:e9a496058fa8b1812bbae4',
-      measurementId: 'G-PKVGVW8D2X',
-    });
-    const messaging = getMessaging(firebaseApp);
-
-    getToken(messaging, {
-      vapidKey:
-        'BFkKBCZ5O4qmyCwm50Aks7sRmMYJzF2wJ8FZCHNLXDLjVxMDEQJFZ_4U5I6uDBF1zXiRHChNAeeDWrTg2m0eL_k',
-    })
-      .then((currentToken) => {
-        if (currentToken) {
-          console.log('currentToken', currentToken);
-          saveToken(currentToken);
-        } else {
-          console.log(
-            'No registration token available. Request permission to generate one.',
-          );
-        }
-      })
-      .catch((err) => {
-        console.log('An error occurred while retrieving token. ', err);
-      });
-
-    onMessage(messaging, (payload) => {
-      console.log('[Foreground]Message received. ', payload);
-      setIsCaptureTime((prev) => !prev);
-    });
-
-    onBackgroundMessage(messaging, (payload) => {
-      console.log(
-        '[firebase-messaging-sw.js] Received background message ',
-        payload,
-      );
-      // });
-
-      const notificationTitle = '[Background] 스크린이 캡처되었습니다!';
-      const notificationOptions = {
-        body: payload,
-        icon: '/public/favicon-32x32.png',
-      };
-
-      self.registration.showNotification(
-        notificationTitle,
-        notificationOptions,
-      );
-    });
-  };
-
   useEffect(() => {
     if (name && room_id) {
       setCurName(name.toString());
@@ -186,10 +107,6 @@ const WebRTCContainer = () => {
       setCoreEndTime();
     }
   }, [curRoomId]);
-
-  useEffect(() => {
-    askPermission();
-  }, []);
 
   return (
     <WebRTCLayout
