+++ package.json
@@ -47,7 +47,7 @@
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev": "next dev -p 4000",
+    "dev": "next dev -p 4001",
     "lint": "next lint",
     "start": "next start"
   },
